# server/build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Tokens": {
				"properties": {
					"access": {
						"type": "string"
					},
					"refresh": {
						"type": "string"
					},
					"tokenVersion": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"access",
					"refresh",
					"tokenVersion"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"enum": [
					"ADMIN",
					"USER",
					"MODERATOR"
				],
				"type": "string"
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"confirmed": {
						"type": "boolean"
					},
					"tokens": {
						"$ref": "#/components/schemas/Tokens"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"id",
					"name",
					"password",
					"confirmed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DishLikes": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"dish_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"user_id",
					"dish_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginInterface": {
				"properties": {
					"token": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"likes": {
						"items": {
							"$ref": "#/components/schemas/DishLikes"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"token",
					"success",
					"likes"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
		"/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"202": {
						"description": "Logged In",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginInterface"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
		"/refresh": {
			"get": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Set-Cookie",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}